+++ packages/tools/src/discolupe/DivContainer.tsx
@@ -1,0 +1,5 @@
+import { DiscoLupeProject } from './data'
+
+export function DivContainer(_: DiscoLupeProject, str: string) {
+  return <div>{str}</div>
+}
