+++ packages/tools/src/discolupe/Popover.tsx
@@ -1,46 +1,46 @@
 // src/Popover.tsx
-import React, { useState, useRef, useEffect, ReactNode } from 'react';
+import React, { useState, useRef, useEffect, ReactNode } from 'react'
 
 interface PopoverProps {
-  content: ReactNode;
-  children: ReactNode;
+  content: ReactNode
+  children: ReactNode
 }
 
 const Popover: React.FC<PopoverProps> = ({ content, children }) => {
-  const [isOpen, setIsOpen] = useState<boolean>(false);
-  const popoverRef = useRef<HTMLDivElement>(null);
-  const triggerRef = useRef<HTMLDivElement>(null);
+  const [isOpen, setIsOpen] = useState<boolean>(false)
+  const popoverRef = useRef<HTMLDivElement>(null)
+  const triggerRef = useRef<HTMLDivElement>(null)
 
   // Toggle popover visibility
   const togglePopover = () => {
-    setIsOpen((prev) => !prev);
-  };
+    setIsOpen((prev) => !prev)
+  }
 
   // Close popover if clicked outside
   const handleClickOutside = (event: MouseEvent) => {
-    const target = event.target as Node;
+    const target = event.target as Node
     if (
       popoverRef.current &&
       !popoverRef.current.contains(target) &&
       triggerRef.current &&
       !triggerRef.current.contains(target)
     ) {
-      setIsOpen(false);
+      setIsOpen(false)
     }
-  };
+  }
 
   useEffect(() => {
     if (isOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
+      document.addEventListener('mousedown', handleClickOutside)
     } else {
-      document.removeEventListener('mousedown', handleClickOutside);
+      document.removeEventListener('mousedown', handleClickOutside)
     }
 
     // Cleanup on unmount
     return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [isOpen]);
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+  }, [isOpen])
 
   return (
     <div className="relative inline-block">
@@ -56,7 +56,7 @@ const Popover: React.FC<PopoverProps> = ({ content, children }) => {
       {isOpen && (
         <div
           ref={popoverRef}
-          className='absolute z-50 mt-2 w-64 rounded-md border border-2 border-orange-500 bg-zinc-950 shadow-lg'
+          className="absolute z-50 mt-2 w-64 rounded-md border border-2 border-orange-500 bg-zinc-950 shadow-lg"
           role="dialog"
           aria-modal="true"
         >
@@ -64,8 +64,7 @@ const Popover: React.FC<PopoverProps> = ({ content, children }) => {
         </div>
       )}
     </div>
-  );
-};
+  )
+}
 
-export default Popover;
-
+export default Popover
